sample-app/pyproject.toml
[project]
name = "google-rmi-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Lepton Software",email = "team@leptonmaps.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn[standard] (>=0.35.0,<0.36.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "brotli-asgi (>=1.4.0,<2.0.0)",
    "google-cloud-bigquery (>=2.0.0)",
    "google-cloud-bigquery-storage (>=2.0.0)",
    "pytz (>=2024.1)",
    "numpy (>=2.3.3,<3.0.0)",
    "pandas (>=2.0.0,<3.0.0)",
    "pyarrow (>=18.0.0,<19.0.0)",
    "db-dtypes (>=1.3.0,<2.0.0)",
    "cachetools (>=5.5.0,<6.0.0)",
    "geojson (>=3.2.0)",
    "shapely",
    "geopandas (>=1.1.1,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
